api: split from __init__ and fix import loops
